Completely changed the logic of exception handling.
Moved the logic of exception handling to the service side. The service sends a generalized BusinessException exception to the controller, which will handle it accordingly. In the controller I created a separate method that will handle the exception and depending on the passed status code the corresponding error page will be displayed.